+++ rororo/openapi/constants.py
@@ -7,11 +7,5 @@ APP_OPENAPI_SCHEMA_KEY = "rororo_openapi_schema"
 #: Key to store OpenAPI spec within the ``web.Application`` instance
 APP_OPENAPI_SPEC_KEY = "rororo_openapi_spec"
 
-#: API Key security scheme
-OPENAPI_SECURITY_API_KEY = "apiKey"
-
-#: HTTP security scheme
-OPENAPI_SECURITY_HTTP = "http"
-
 #: Key to store valid OpenAPI context within the ``web.Request`` instance
 REQUEST_OPENAPI_CONTEXT_KEY = "rororo_openapi_context"