+++ rororo/openapi/validators.py
@@ -1,80 +1,91 @@
 from typing import Any, List, Tuple
 
+import pyrsistent
+from email_validator import EmailNotValidError, validate_email
+from jsonschema.exceptions import FormatError
+from openapi_core.exceptions import OpenAPIError
 from openapi_core.schema.operations.models import Operation
 from openapi_core.schema.specs.models import Spec
+from openapi_core.unmarshalling.schemas.formatters import Formatter
+from openapi_core.validation.request.datatypes import (
+    OpenAPIRequest,
+    RequestParameters,
+)
 from openapi_core.validation.request.validators import (
     RequestValidator as BaseRequestValidator,
 )
+from openapi_core.validation.response.datatypes import OpenAPIResponse
 from openapi_core.validation.response.validators import ResponseValidator
 
-from .data import (
-    OpenAPICoreRequest,
-    OpenAPICoreResponse,
-    OpenAPIParameters,
-    to_openapi_parameters,
-)
+from .data import OpenAPIParameters, to_openapi_parameters
 from .exceptions import ValidationError
-from .mappings import enforce_dicts, enforce_immutable_data, merge_data
+from .security import validate_security
+from .utils import get_base_url
+from ..annotations import MappingStrAny
 
 
-CUSTOM_FORMATTERS = {"email": str}
-ROOT_PATH = "."
+class EmailFormatter(Formatter):
+    def validate(self, value: str) -> bool:
+        try:
+            validate_email(value)
+        except EmailNotValidError as err:
+            raise FormatError(f"{value!r} is not an 'email'", cause=err)
+        return True
+
+
+CUSTOM_FORMATTERS = {"email": EmailFormatter()}
 
 
 class RequestValidator(BaseRequestValidator):
-    """Custom request validator for rororo.
+    def _get_parameters(
+        self, request: OpenAPIRequest, params: MappingStrAny
+    ) -> Tuple[RequestParameters, List[OpenAPIError]]:
+        parameters, errors = super()._get_parameters(request, params)
+        if errors:
+            raise ValidationError.from_request_errors(
+                errors, unmarshal_loc=["parameters"]
+            )
+        return parameters, errors
 
-    By default, openapi-core doesn't support proper assigining additional
-    properties into the resulted body model, cause of that instead of returning
-    Models instances - return casted, but unmarshaled request body data.
-    """
-
-    def _get_body(
-        self, request: OpenAPICoreRequest, operation: Operation
-    ) -> Tuple[Any, List[Exception]]:
-        body, errors = super()._get_body(request, operation)
-        if errors or not operation.request_body:
-            return (body, errors)
-
-        body_with_dicts = enforce_dicts(body)
-
-        media_type = operation.request_body[request.mimetype]
-        raw_body = operation.request_body.get_value(request)
-
-        return (
-            enforce_immutable_data(
-                merge_data(body_with_dicts, media_type.cast(raw_body))
-            ),
-            errors,
-        )
+    def _get_security(
+        self, request: OpenAPIRequest, operation: Operation
+    ) -> MappingStrAny:
+        return validate_security(self, request, operation)
 
 
-def validate_request_parameters_and_body(
-    spec: Spec, core_request: OpenAPICoreRequest
-) -> Tuple[OpenAPIParameters, Any]:
+def validate_request(
+    spec: Spec, core_request: OpenAPIRequest
+) -> Tuple[Any, OpenAPIParameters, Any]:
     """
     Instead of validating request parameters & body in two calls, validate them
     at once with passing custom formatters.
     """
-    validator = RequestValidator(spec, custom_formatters=CUSTOM_FORMATTERS)
+    validator = RequestValidator(
+        spec,
+        custom_formatters=CUSTOM_FORMATTERS,
+        base_url=get_base_url(core_request),
+    )
     result = validator.validate(core_request)
 
     if result.errors:
         raise ValidationError.from_request_errors(result.errors)
 
     return (
+        result.security,
         to_openapi_parameters(result.parameters),
-        result.body,
+        pyrsistent.freeze(result.body),
     )
 
 
-def validate_response_data(
-    spec: Spec,
-    core_request: OpenAPICoreRequest,
-    core_response: OpenAPICoreResponse,
+def validate_response(
+    spec: Spec, core_request: OpenAPIRequest, core_response: OpenAPIResponse,
 ) -> Any:
     """Pass custom formatters on validating response data."""
-    validator = ResponseValidator(spec, custom_formatters=CUSTOM_FORMATTERS)
+    validator = ResponseValidator(
+        spec,
+        custom_formatters=CUSTOM_FORMATTERS,
+        base_url=get_base_url(core_request),
+    )
     result = validator.validate(core_request, core_response)
 
     if result.errors: